Run the webpack compilation in worker threads
