+++ package.json
@@ -148,6 +148,7 @@
     "mocha": "^9.1.3",
     "node-fetch": "^2.6.6",
     "nyc": "^15.1.0",
+    "piscina": "^3.2.0",
     "playwright": "^1.17.1",
     "prettier": "^2.5.1",
     "process": "^0.11.10",
